refactor: add modules, c/s architecture
