+++ controllers/inputController.js
@@ -1,0 +1,17 @@
+export function inputHandle (input) {
+  const formattedInput = input.trim();
+  switch (formattedInput.split(' ')[0]) {
+    case 'up': {
+      console.log('up!');
+      // currentPath = currPathUp();
+      break;
+    }
+    case 'cd': {
+      // currentPath = currPathCd(input);
+      break;
+    }
+    default:
+      console.log('Invalid input');
+      break;
+  }
+}
